+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 #### 11.09.2018
 
+* [Android] Visualize: Edit Text Preference Constraints
 * [Android] Visualize: Edit Text Preference
 
 #### 10.09.2018